- APDS99xx: use of config bits (instead of Status register ones) to trigger proper interrupt(s) clear function - PCA9685: Dedicated functions to compute PWM register values (useful to manually fill an array with registers values to send multiple channels at once in one I2C transaction) - shared_CLS: double precision now single precision - README.md update

+++ APDS9960_proc.h
@@ -50,7 +50,7 @@ typedef struct APDS9960_t {
 	bool						Saturation;			//!< Current Sensor saturation status
 	struct {
 	I2C_slave_t *				slave_inst;			//!< Slave structure
-	double						mat[3][3];			//!< RGB to XYZ matrix
+	float						mat[3][3];			//!< RGB to XYZ matrix
 	APDS9960_als_gain			ADC_Gain;			//!< ADC (ALS & Color) sensor gain config
 	APDS9960_prox_gain			Prox_Gain;			//!< Proximity sensor gain config
 	APDS9960_pulse_length		Prox_Length;			//!< Proximity LED pulse length
